Change Projects from data to collection
Because we're displaying the resume on the landing page with links to
each project, I think using a collection for the project makes the most
sense, so we can have a page for each project with more details if
someone wants to click through.

I've also set up the layout to be responsive and be column based on
mobile vs sidebar on larger screens.